requirements.txt
about-time==3.1.1
absl-py==2.2.2
addict==2.4.0
adjustText==0.8
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.3.1
alive-progress==2.2.0
anyio==3.5.0
appdirs==1.4.4
apptools==5.3.0
asttokens==2.4.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attr==0.3.2
attributee==0.1.8
attrs==23.2.0
backcall==0.2.0
blessed==1.20.0
blinker==1.8.2
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
build==1.2.2
cachetools==4.2.4
calmsize==0.1.3
ccimport==0.3.7
certifi @ file:///croot/certifi_1700501669400/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
cftime==1.6.0
charset-normalizer==3.4.2
cityscapesScripts==2.2.4
click @ file:///croot/click_1698129812380/work
cloudpickle==3.0.0
cmake==4.0.2
codecov==2.1.13
colorama==0.4.4
coloredlogs==15.0.1
comet-ml==3.31.21
comm==0.2.2
commonmark==0.9.1
ConfigArgParse==1.7
configobj==5.0.6
contourpy==1.1.1
coverage==7.4.0
crc32c==2.2.post0
cryptography @ file:///croot/cryptography_1694444244250/work
csaps==1.1.0
cumm-cu111==0.2.8
cupy-cuda111==10.2.0
cycler==0.12.1
Cython==0.29.20
cytoolz @ file:///croot/cytoolz_1667465931118/work
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask @ file:///croot/dask-core_1683065217061/work
dataclasses==0.6
decorator==5.2.1
deepspeed==0.6.5
defusedxml==0.7.1
deprecation==2.1.0
diffusers==0.11.1
docker-pycreds==0.4.0
dominate==2.9.1
dotty-dict==1.3.1
drjit==0.2.1
dulwich==0.20.32
easydict==1.10
einops==0.8.1
envisage==7.0.3
everett==3.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
fastrlock==0.8
filelock==3.13.1
fire==0.4.0
flake8==6.1.0
Flask==3.0.3
flatbuffers==2.0
flatten-dict==0.4.2
fonttools==4.29.1
freetype-py==2.3.0
frozenlist==1.3.0
fsspec==2025.3.0
ftfy==6.0.3
future==0.18.2
fvcore==0.1.5.post20220512
gast==0.5.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpustat==1.1.1
grapheme==0.6.0
grpcio==1.70.0
hjson==3.0.2
huggingface-hub==0.30.2
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
imagecodecs @ file:///croot/imagecodecs_1695064943445/work
imagecorruptions==1.1.2
imageio==2.35.1
imageio-ffmpeg==0.4.5
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
interrogate==1.5.0
iopath==0.1.10
ipython==8.11.0
isort==4.3.21
itsdangerous==2.2.0
jedi==0.19.2
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpath-rw==1.4.0
jsonpointer==2.4
jupyter-packaging==0.12.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kornia==0.6.6
kwarray==0.6.16
lap==0.4.0
lark==1.1.2
lazy_loader==0.4
libclang==14.0.1
lit==18.1.8
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
loguru==0.7.2
Markdown==3.7
mat4py==0.6.0
matplotlib==3.7.5
matplotlib-inline==0.1.7
matplotlib2tikz==0.7.6
mayavi==4.8.2
mccabe==0.7.0
mdurl==0.1.2
meshio==5.3.4
mitsuba==3.0.1
mkl-fft @ file:///croot/mkl_fft_1695058164594/work
mkl-random @ file:///croot/mkl_random_1695059800811/work
mkl-service==2.4.0
mlcroissant==1.0.5
mmcls==0.25.0
mmcv==1.3.0
mmdet==3.3.0
mmengine==0.10.2
mmpretrain==1.2.0
model-index==0.1.11
modelindex==0.0.2
monai==1.3.0
motmetrics==1.4.0
mpmath==1.3.0
mrcfile==1.3.0
multidict==6.0.2
multipledispatch==0.6.0
MultiScaleDeformableAttention==1.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nestargs==0.5.0
netCDF4==1.5.8
networkx @ file:///croot/networkx_1690561992265/work
ninja==1.10.2.3
numpy==1.22.4
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.7.0
onnxruntime==1.16.3
open-clip-torch==2.24.0
opencv-python==4.11.0.86
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
OpenEXR==1.3.7
openmim==0.3.9
openxlab==0.0.11
opt-einsum==3.3.0
ordered-set==4.1.0
packaging @ file:///croot/packaging_1693575174725/work
pandas==1.3.5
pandas-stubs==2.0.3.230814
parso==0.8.4
partd @ file:///croot/partd_1698702562572/work
pathtools==0.1.2
pccm==0.3.4
pickleshare==0.7.5
Pillow==9.4.0
pip-tools==7.4.1
platformdirs==4.3.6
plotly==5.24.1
pluggy==1.5.0
ply==3.11
plyfile==0.7.4
pooch @ file:///croot/pooch_1695850093751/work
portalocker==2.5.1
prettytable==3.11.0
progressbar2==4.0.0
promise==2.3
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptflops==0.7.3
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.1
pyDeprecate==0.3.1
pyface==8.0.0
pyflakes==3.1.0
pyglet==1.5.23
Pygments==2.19.1
pykeops==1.5
PyMCubes==0.1.2
PyOpenGL==3.1.0
pyOpenSSL @ file:///croot/pyopenssl_1690223430423/work
pyparsing==3.1.4
pyproject_hooks==1.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyquaternion==0.9.9
pyrr==0.10.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pysodmetrics==1.4.2
pytest==8.3.3
python-box==6.1.0
python-dateutil==2.9.0.post0
python-swiftclient==4.0.0
python-utils==3.3.3
pytorch-lightning==1.5.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2020.10.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
retrying==1.3.4
rich==13.1.0
rsa==4.9.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.11.2
selective_scan==0.0.2
semantic-version==2.9.0
sentencepiece==0.2.0
setproctitle==1.3.3
shapely==2.0.6
SharedArray==3.2.1
simple-parsing==0.0.18
simplejson==3.18.0
simplification==0.7.10
six==1.16.0
sniffio==1.2.0
stack-data==0.6.3
submitit==1.5.2
supervision==0.25.1
sympy==1.13.3
tabulate==0.8.9
tenacity==9.0.0
tensorboard==2.2.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6
tensorflow-gpu==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminaltables==3.1.10
tf-estimator-nightly==2.8.0.dev2021122109
tflearn==0.5.0
tfrecord==1.14.1
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2023.7.10
tikzplotlib==0.10.1
timm==0.3.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.10.0
toolz @ file:///croot/toolz_1667464077321/work
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.7.2
torchsummary==1.5.1
torchvision==0.12.0
tornado==6.4
tqdm==4.51.0
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
transformers==4.36.2
transforms3d==0.4.2
trimesh==3.10.1
types-pytz==2024.2.0.20241221
typing==3.7.4.3
typing-inspect==0.7.1
typing_extensions @ file:///croot/typing_extensions_1690297465030/work
tzdata==2023.3
ubelt==1.3.4
urllib3==2.2.3
visdom==0.2.4
vtk==9.3.1
wcwidth==0.2.13
websocket-client==1.7.0
Werkzeug==3.0.6
wrapt==1.13.3
wurlitzer==3.0.2
xdoctest==1.1.2
xmltodict==0.13.0
yacs==0.1.8
yapf==0.40.2
yarl==1.7.2
yaspin==2.1.0
zipp==3.20.2
